+++ spec/dummy/config/routes.rb
@@ -1,0 +1,5 @@
+Gigs::Application.routes.draw do
+  resources :concerts
+
+  get 'locations/:location/concerts', to: 'concerts#index'
+end