+++ spec/spec_helper.rb
@@ -1,0 +1,4 @@
+RSpec.configure do |config|
+  config.filter_run_excluding failing: true
+  config.filter_run_excluding wip: true
+end