+++ HISTORY.md
@@ -1,3 +1,7 @@
+v0.2.0  - 2013/10/20 Include specs for local and remote -- might break
+                     compatibility since it requires both dependencies in the
+                     Gemfile (e.g., might have conflicts with Rails 3)
+
 v0.1.2  - 2013/10/14 Support JSONP callbacks
 
 v0.1.1  - 2013/10/10 Add PATCH, HTTP Caching, :comparing_with for filters